+++ example-vite-antd/src/views/Dashboard.vue
@@ -1,0 +1,11 @@
+<template>
+  <h1>Dashboard</h1>
+</template>
+<script>
+export default {
+  setup() {
+    return {
+    }
+  }
+}
+</script>